requirements.txt
anthropic>=0.39.0
tree-sitter>=0.23.0
tree-sitter-java>=0.23.4
lxml>=5.1.0
requests>=2.31.0
python-dotenv>=1.0.0
